requirements.txt
alembic==1.7.5
anyio==3.4.0
asgiref==3.4.1
asyncpg==0.25.0
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
Deprecated==1.2.13
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
Mako==1.1.6
MarkupSafe==2.0.1
mypy==0.910
mypy-extensions==0.4.3
passlib==1.7.4
pika==1.2.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
PyMySQL==1.0.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
redis==4.0.2
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.28
starlette==0.16.0
toml==0.10.2
typing_extensions==4.0.1
uvicorn==0.16.0
wrapt==1.13.3
